enemies spawn, player can shoot and kill

+++ Source/Engine/Renderer/Text.cpp
@@ -31,7 +31,7 @@ namespace parabellum {
 		return true;
 	}
 
-	void Text::Draw(Renderer& renderer, int x, int y)  {
+	void Text::Draw(Renderer& renderer, float x, float y)  {
 		// get the texture width and height
 		float width, height;
 		bool success = SDL_GetTextureSize(m_texture, &width, &height);
